ui/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==9.5.0
torch==2.0.1
transformers==4.30.2 